chore: rm chembench from corral dep and add to env
