+++ pyproject.toml
@@ -46,7 +46,6 @@ dependencies = [
   "modal==0.70.3",
   "rich>=13.9.4",
   "ipython>=8.18.1",
-  "chembench @ git+https://github.com/lamalab-org/chembench.git@main",
   "promptstore @ git+https://github.com/lamalab-org/promptstore.git@main",
   "datasets==2.19.1",
   "huggingface-hub==0.26.0",
